No need to reference the icon in Tray
